+++ lib/meta_tags/renderer.rb
@@ -51,8 +51,10 @@ module MetaTags
     # @param [Array<Tag>] tags a buffer object to store tag in.
     #
     def render_title(tags)
+      normalized_meta_tags[:title] = meta_tags.page_title
+      normalized_meta_tags[:site] = meta_tags[:site]
       title = meta_tags.extract_full_title
-      normalized_meta_tags[:title] = title
+      normalized_meta_tags[:full_title] = title
       tags << ContentTag.new(:title, content: title) if title.present?
     end