+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Features:
 
 Bugfixes:
   - When title limit reached with `reverse` set to `true`, properly truncate the last item of the title array instead of the first one.
+  - Do not merge title and site title for OpenGraph, site title is available for reference as `:site`, and full title as `:full_title`.
 
 Changes:
   - Removed Google "author" and "publisher" links, as Google deprecated these options (https://support.google.com/webmasters/answer/6083347?hl=en)